# Conflicts:
#	src/main/java/game/model/SentPlayerData.java
#	src/main/resources/static/js/game.js

+++ src/main/java/game/config/constant/ShieldConfig.java
@@ -1,8 +1,8 @@
 package game.config.constant;
 
 public class ShieldConfig {
-    public final static long NO_SHIELD_PROTECTION = 0L;
-    public final static long NORMAL_SHIELD_PROTECTION = 10L;
-    public final static long PLASMA_SHIELD_PROTECTION = 15L;
-    public final static long ATOM_SHIELD_PROTECTION = 20L;
+    public final static double NO_SHIELD_PROTECTION = 0L;
+    public final static double NORMAL_SHIELD_PROTECTION = 10.0;
+    public final static double PLASMA_SHIELD_PROTECTION = 15.0;
+    public final static double ATOM_SHIELD_PROTECTION = 20.0;
 }
